+++ CHANGES.rst
@@ -19,6 +19,14 @@
 Changes
 =======
 
+Version 0.2.0 (2025-04-02)
+--------------------------
+- Added get_collections Example - Included a practical usage example for the get_collections function.
+- Reworked Plot Function Using Plotly - Updated the plotting function to Plotly, instead of matplotlib, for interactive visualizations.
+- Reworked get_phenometrics_region Function - List phenological metrics calculated for each spatial location of the given region.
+- Added get_timeseries_region Function - Retrieves the satellite images time series for each pixel centers within the given region.
+- Remade Jupyter Notebooks - Updated and restructured Jupyter notebooks to reflect new functionalities and a step-by-step guidance.
+- Updated Documentation - Revised and expanded documentation to include new features and improve readability.
 
 Version 0.1.0 (2025-03-10)
 --------------------------
@@ -27,5 +35,5 @@ Version 0.1.0 (2025-03-10)
 - Add cube_query - An object that contains the information associated with a collection that can be downloaded or acessed.
 - Add get_collections - List available data cubes in the BDC's SpatioTemporal Asset Catalogs (STAC).
 - Add get_description - List the information on each of the phenological metrics, such as code, name, description and method.
-- Add get_phenometrics_region - List phenological metrics calculated for each of the given spatial location based on selected region methodology (all, systematic or random grid).
+- Add get_phenometrics_region - List phenological metrics calculated for each of the given spatial location based on selected methodology.